The code decompiled, with main identified
What is here is a decompilation from Ghidra. To do this, it was not
possible to simply decompile the steam version of swep1r, because of
some sort of encryption. So instead, I ran the game and created a memory
dump of the game. That was too large ~400MB for Ghidra to handle well.
So I dumped out the binary contents of the executable one segment at a
time - i.e. header, .text, .rdata, .data, .bind etc

After this, I concatenated those binary files together, and then used
that within the Ghidra analysis, which worked much better